+++ assets/ComponentsPage-4be1d88d.js
@@ -1,0 +1,23 @@
+var m=Object.defineProperty;var e=(n,r)=>m(n,"name",{value:r,configurable:!0});import{C as p}from"./ComponentInfo-1eac72a6.js";import{d as a}from"./ListItemText-be55e132.js";import{l as i,c as t}from"./ErrorPage-eb59dadb.js";import{T as s}from"./ThemeProvider-0a69d799.js";import"./Link-123f7416.js";import"./copyText-dae2e1ef.js";import"./IconButton-c6b31999.js";import"./Paper-32f89a64.js";import"./PaperCode-b3aeccfd.js";import"./PageNav-652b02ba.js";import"./_commonjsHelpers-0119a68f.js";import"./sdk.m-5f3892a8.js";import"./vite.config.ts-faa4c0cf.js";const l=i({components:{MuiButtonBase:{defaultProps:{disableRipple:!0}}}});function o(){return t(s,{theme:l,get children(){return t(a,{children:"Change default props"})}})}e(o,"DefaultPropsExample");o.code=`import { Button } from "@suid/material";
+import { createTheme, ThemeProvider } from "@suid/material/styles";
+
+const theme = createTheme({
+  components: {
+    // Name of the component ⚛️
+    MuiButtonBase: {
+      defaultProps: {
+        // The default props to change
+        disableRipple: true, // No more ripple, on the whole application 💣!
+      },
+    },
+  },
+});
+
+export default function DefaultPropsExample() {
+  return (
+    <ThemeProvider theme={theme}>
+      <Button>Change default props</Button>
+    </ThemeProvider>
+  );
+}
+`;function b(){return t(p,{name:"Components",importInfo:!1,moreExamples:!1,docsApiName:"customization/components",examples:[o]})}e(b,"StyledPage");export{b as default};